move classes for "create solution" to a new package
