+++ ui/src/main/java/de/ipb_halle/lbac/items/bean/aliquot/createsolution/consumepartofitem/ConsumePartOfItemStrategyController.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  */
-package de.ipb_halle.lbac.items.bean.createsolution.consumepartofitem;
+package de.ipb_halle.lbac.items.bean.aliquot.createsolution.consumepartofitem;
 
 import java.io.Serializable;
 import java.util.Date;
