+++ ui/src/test/java/de/ipb_halle/lbac/items/bean/aliquot/createsolution/consumepartofitem/ConsumePartOfItemStep4ControllerTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  */
-package de.ipb_halle.lbac.items.bean.createsolution.consumepartofitem;
+package de.ipb_halle.lbac.items.bean.aliquot.createsolution.consumepartofitem;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
